Create scanner
File create a dictionary by scanning all files in a given directory (Skyrim Special Edition folder) and get what plugin is considered a 'master' of the file. For example, it will get what plugin names that are in a _DISTRI.ini (PowerOfThree's Spell Perk Item Distributor ini file) and add the file as a new value to a key which is a plugin name.