Added methods to process NVPP data separately.
Created save_nvpp_data() which is called after decompressing all forms. This method separates the NVPP field from the raw data so that when re.sub is called during compacting/patching. the data of the NVPP field is not mistakenly replaced.
Added fix_nvpp_data() method to process the NVPP data separately due to it's structure.
Removed '|' pipes from regex expressions where it was mistakenly included.
Changed ~||+||~ flags so that on consecutive replace, new form ids are not replaced again.