Optimization. 20+ times faster.
Removed the data_list re.split as it was much too slow. Instead opted to use str.count() in conjunction with str.index() as we only need to care about b'CELL' presence in the file instead of all form types.

Removed threading as the script is so fast it is unnecessary.