overhaul compacting/dependent patching + bug fixes
Import the regex library as re instead of using python's included re library.
We no longer regex over the entire file and instead search for the master byte in the form ID to get all possible form ids. We then iterate over those with conditions to check if it is a form id and if it needs patching (~5x faster).
Changed data to a bytearray so that we don't need to deal with the immutability of bytes.
Added a method to properly patch LAND records.
Change the data_list split method to use the actual record types instead of regex matching [A-Z]{3}[A-Z_].
Fixed issue where when using MO2 mode certain files were not added to the comapacted_and_patched dictionary.
Fixed an issue where get_master_index's regex pattern would not match certain plugin names by adding the re.DOTALL flag.
Fixed an issue where some GRUPs would match the conditions to get through the decompress and recompress methods.
Moved some things to accommodate the new get_land_data and patch_land_data methods.