Create full_form_processor.py
save_all_form_data() iterates through data_list and for each type of record sends it to the relevant method. Each save_FORM_data method returns a list of [index, bytearray(form), FORM_offsets] where the index is the order the form is in the file, the bytearray(form) is the entire form raw data in bytearray for more efficient patching and the FORM_offsets is a list of byte offsets for that form where each offset is the location of a form id.

patch_form_data() takes the saved_forms from save_all_form_data and patches the relevant form ids in each form and reassembles the data_list.

vmad_reader() currently takes master_byte as a parameter but doesn't use it. May use it to patch script file names if necessary but it seems like it might not be.