Dynamic Animation Casting toml patching + bug fix
Fix issue in ini_mu_patcher that would occur when the new form id was much shorter in character length then the old one.

Added a method of patching toml file for Dynamic Animation Casting. Patching the NG version's toml files is more reliable. The old version of the toml file patching assumes that the conditions are in the exact same line order as the old template and no extra white spaces.