Major optimization of getting grup_struct()
import intervaltree and use it to calculate the grup_struct() this is much faster than the original method.
Remove the printing for the sizes which gave a general estimate of time it would take to patch since we are patching much faster now. Thread every dependent patch as their is no point not to,