Fix issue for plugins without Skyrim.esm as master
Plugins without skyrim.esm as master need to start their Form IDs at 0x800 regardless of whether or not they have 1.71 header.
Plugins with dependents that do not have Skyrim.esm as master also need have their Form IDs start at 0x800.

These are because if a plugin is flagged ESL without Skyrm.esm as a master then Form IDs below 0x800 completely overwrite Skyrim.esm records. The same goes for if a plugin's dependents. If the master has Form IDs below 0x800 and the dependent is missing Skyrim.esm as a master then the dependent loses track of the master at position 0 and assumes it is supposed to be changing Skyrim.esm records.